% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{metro4bgcolors}
\alias{metro4bgcolors}
\alias{metro4fgcolors}
\title{Metro 4 colors}
\usage{
metro4bgcolors()

metro4fgcolors()
}
\value{
The vector of available colors.
}
\description{
Returns the vector of available colors.
}
