% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{multiActionButton}
\alias{multiActionButton}
\title{Multi action button}
\usage{
multiActionButton(
  rotate = TRUE,
  bg = NULL,
  fg = NULL,
  icon = NULL,
  direction = "right",
  subButtons
)
}
\arguments{
\item{rotate}{logical, whether to rotate the icon upon clicking the main
button}

\item{bg}{background color of the button}

\item{fg}{color of the icon}

\item{icon}{icon name}

\item{direction}{the opening direction, \code{"top"}, \code{"right"},
\code{"bottom"}, or \code{"left"}}

\item{subButtons}{a list of buttons created with \code{\link{subButton}}}
}
\description{
Creates a multi action button for usage in a Shiny UI.
}
\examples{
#XXX
}
