% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{shinyMetro4bgColors}
\alias{shinyMetro4bgColors}
\title{Shiny app displaying the metro 4 background colors}
\usage{
shinyMetro4bgColors()
}
\description{
Launch a shiny app showing all available metro 4 background colors.
}
