% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subButton}
\alias{subButton}
\title{Sub-button}
\usage{
subButton(
  id,
  bg = NULL,
  fg = NULL,
  icon = NULL,
  iconSize = "3x",
  tooltip = NULL,
  onclick = NULL
)
}
\arguments{
\item{id}{button id}

\item{bg}{background color of the button}

\item{fg}{color of the icon}

\item{icon}{icon name}

\item{iconSize}{icon size, can be \code{"lg"}, \code{"2x"}, \code{"3x"},
\code{"4x"} or \code{"5x"}}

\item{tooltip}{character string to show as tooltip}

\item{onclick}{\code{onclick} attribute}
}
\description{
Creates a sub-button for usage in \code{\link{multiActionButton}}
}
